Archivo binario "codecept-skeleton" añadido. Clase de instalador aislada para un mejor control.
