+++ composer.json
@@ -14,10 +14,12 @@
         "guzzlehttp/guzzle": ">=4.1.4 <7.0"
     },
     "autoload": {
+        "psr-4": { "CodeceptionSkeleton\\": "src/" },
         "classmap": [
             "ComposerListener.php"
         ]
     },
+    "bin": ["codecept-skeleton"],
     "scripts": {
         "deploy": [
             "ComposerListener::downloadDrivers",
